convert credits and credit_cost db columns to decimal

+++ db/migrate/20160328220449_add_credit_cost_to_courses.rb
@@ -1,5 +1,5 @@
 class AddCreditCostToCourses < ActiveRecord::Migration
   def change
-    add_column :courses, :credit_cost, :integer
+    add_column :courses, :credit_cost, :decimal
   end
 end
